test: move API tests to Feature/API
- Moved `PlanControllerTest`, `SubscriptionMethodControllerTest`, `SubscriptionControllerTest`, and `SubscriptionHistoryControllerTest` from `tests/Http/Controllers/API` to `tests/Feature/API`.
- Updated namespaces of the test files to `Juzaweb\Modules\Subscription\Tests\Feature\API`.
